requirements.txt
librosa==0.10.1
soundfile
moviepy
gtts
SpeechRecognition
streamlit
regex
google-generativeai
python-dotenv